backend/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mypregnancy-api"
version = "0.1.0"
description = "The FastAPI backend for the 'MyPregnancy' mobile app and dashboard web-app"
requires-python = ">=3.12"

# -------------------------------------------
# -------- PRODUCTION DEPENDENCIES ----------
# -------------------------------------------
dependencies = [
    "itsdangerous>=2.2.0",
    "fastapi>=0.121.2",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.44",
    "alembic>=1.17.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "argon2-cffi>=25.1.0",
    "pwdlib>=0.3.0",
    "boto3>=1.40.74",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "pyjwt>=2.10.1",
    "authlib>=1.6.5",
    "loguru>=0.7.3",
    "sentry-sdk>=2.44.0",
    "beautifulsoup4>=4.14.2",
    "requests>=2.32.5",
    "asyncpg>=0.31.0",
]

# --------------------------------------------
# -------- DEVELOPMENT DEPENDENCIES ----------
# --------------------------------------------
[dependency-groups]
dev = [
    "boto3-stubs>=1.40.74",
    "botocore-stubs>=1.40.74",
    "faker>=38.0.0",
    "fastapi-cli>=0.0.16",
    "httpx>=0.28.1",
    "ipython>=9.7.0",
    "jupyterlab>=4.4.10",
    "mypy-boto3-s3>=1.40.61",
    "pytest>=9.0.1",
]

# --- ALL [tool.*] SECTIONS GO AT THE END ---
[tool.ruff]
line-length = 120

[tool.setuptools.packages.find]
include = ["app", "app.*"]
